Grid and particle generation

Tesserae.generate_grid — Function

generate_grid(GridProp, mesh)

Generate a background grid with the type GridProp. This returns a StructArray (StructArrays.jl). The first field of GridProp is used to store mesh, which requires Vec{dim, T} == eltype(mesh).

Examples

julia> struct GridProp{dim, T}
           x  :: Vec{dim, T}
           m  :: Float64
           mv :: Vec{dim, T}
           f  :: Vec{dim, T}
           v  :: Vec{dim, T}
           vⁿ :: Vec{dim, T}
       end

julia> grid = generate_grid(GridProp{2, Float64}, CartesianMesh(0.5, (0,3), (0,2)));

julia> grid[1]
GridProp{2, Float64}([0.0, 0.0], 0.0, [0.0, 0.0], [0.0, 0.0], [0.0, 0.0], [0.0, 0.0])

julia> grid.x
7×5 CartesianMesh{2, Float64, Vector{Float64}}:
 [0.0, 0.0]  [0.0, 0.5]  [0.0, 1.0]  [0.0, 1.5]  [0.0, 2.0]
 [0.5, 0.0]  [0.5, 0.5]  [0.5, 1.0]  [0.5, 1.5]  [0.5, 2.0]
 [1.0, 0.0]  [1.0, 0.5]  [1.0, 1.0]  [1.0, 1.5]  [1.0, 2.0]
 [1.5, 0.0]  [1.5, 0.5]  [1.5, 1.0]  [1.5, 1.5]  [1.5, 2.0]
 [2.0, 0.0]  [2.0, 0.5]  [2.0, 1.0]  [2.0, 1.5]  [2.0, 2.0]
 [2.5, 0.0]  [2.5, 0.5]  [2.5, 1.0]  [2.5, 1.5]  [2.5, 2.0]
 [3.0, 0.0]  [3.0, 0.5]  [3.0, 1.0]  [3.0, 1.5]  [3.0, 2.0]

julia> grid.v
7×5 Matrix{Vec{2, Float64}}:
 [0.0, 0.0]  [0.0, 0.0]  [0.0, 0.0]  [0.0, 0.0]  [0.0, 0.0]
 [0.0, 0.0]  [0.0, 0.0]  [0.0, 0.0]  [0.0, 0.0]  [0.0, 0.0]
 [0.0, 0.0]  [0.0, 0.0]  [0.0, 0.0]  [0.0, 0.0]  [0.0, 0.0]
 [0.0, 0.0]  [0.0, 0.0]  [0.0, 0.0]  [0.0, 0.0]  [0.0, 0.0]
 [0.0, 0.0]  [0.0, 0.0]  [0.0, 0.0]  [0.0, 0.0]  [0.0, 0.0]
 [0.0, 0.0]  [0.0, 0.0]  [0.0, 0.0]  [0.0, 0.0]  [0.0, 0.0]
 [0.0, 0.0]  [0.0, 0.0]  [0.0, 0.0]  [0.0, 0.0]  [0.0, 0.0]

Tesserae.generate_particles — Function

generate_particles([ParticleProp], mesh; spacing=1/2, alg=PoissonDiskSampling())

Generate particles across the entire mesh domain with a particle spacing (0 < spacing ≤ 1). When using GridSampling(), setting spacing = 1/η will produce η^dim particles per cell, where dim is the problem dimension. With PoissonDiskSampling(), a similar number of particles will be generated as with GridSampling().

julia> mesh = CartesianMesh(0.5, (0,3), (0,2));

julia> generate_particles(mesh; alg=GridSampling())
96-element Vector{Vec{2, Float64}}:
 [0.125, 0.125]
 [0.375, 0.125]
 [0.625, 0.125]
 [0.875, 0.125]
 [1.125, 0.125]
 [1.375, 0.125]
 [1.625, 0.125]
 [1.875, 0.125]
 [2.125, 0.125]
 [2.375, 0.125]
 ⋮
 [0.875, 1.875]
 [1.125, 1.875]
 [1.375, 1.875]
 [1.625, 1.875]
 [1.875, 1.875]
 [2.125, 1.875]
 [2.375, 1.875]
 [2.625, 1.875]
 [2.875, 1.875]

julia> generate_particles(mesh; alg=PoissonDiskSampling())
104-element Vector{Vec{2, Float64}}:
 [0.019486302766976688, 0.10084456001030316]
 [0.3706080641664693, 0.007791503721806592]
 [0.5719917675750236, 0.05464808529241699]
 [0.8133929361759007, 0.041653701239464935]
 [1.2018738953580164, 0.01162088274898504]
 [1.4828122557187768, 0.11160372106598743]
 [1.7056259797447908, 0.12391213709761462]
 [1.9797445655479007, 0.014712577869719523]
 [2.278492717206701, 0.12390578866223867]
 [2.546975361348073, 0.0031820905928220955]
 ⋮
 [0.4898559143708251, 1.89356020600053]
 [1.0195810307629698, 1.889052302879482]
 [1.9966812579524504, 1.9495828742294676]
 [2.227550017652492, 1.9511715730143793]
 [2.5110348565493656, 1.960703712587761]
 [2.7242620130391613, 1.8586667688178318]
 [0.6667104926512478, 1.996534016067242]
 [1.4028610486139148, 1.9964831612157148]
 [1.7433278603215723, 1.9817737591842066]